iox-#1032 Refactor error reporting
