+++ iceoryx_hoofs/include/iceoryx_hoofs/error_handling_2/platform/error_reporting_dynamic.hpp
@@ -40,7 +40,6 @@ struct Handler
     }
 };
 
-// set all pointers or nothing (whole handler)
 // could also be a singleton etc.
 Handler g_handler;
